np.int -> np.int_ and np.float -> np.float_

+++ sub-sources-outside-region.py
@@ -100,11 +100,11 @@ def getimsize(image):
     his = hdul[0].header['HISTORY']
     for line in his:
         if 'Image-NPix' in line:
-            imsizeddf = np.int(line.split('=')[1])
+            imsizeddf = np.int_(line.split('=')[1])
         elif 'Image-Cell' in line:
-            cellddf = np.float(line.split('=')[1])
+            cellddf = np.float_(line.split('=')[1])
         elif 'Weight-Robust' in line:
-            robustddf = np.float(line.split('=')[1])
+            robustddf = np.float_(line.split('=')[1])
     
 
     if imsizeddf is None:
@@ -1091,7 +1091,7 @@ for observation in range(number_of_unique_obsids(msfiles)):
 
     if split:
 
-        nchanperblock = np.int(20/freqstepavg)
+        nchanperblock = np.int_(20/freqstepavg)
         t = pt.table(currentmsoutconcat + '/SPECTRAL_WINDOW', readonly=True)
         nchan = t.getcol('NUM_CHAN')[0]
         t.close()
